Don't open a new terminal when synthesizing
Instead, re-use `stdout` in the current terminal. We do this because
not all machines have `gnome-terminal` installed, nor is it convenient
to have to close the spawned instances of `gnome-terminal` each time
you synthesize your piece.

This is also less prone to breaking; `gnome-terminal` has deprecated
the -e flag and that behavior is not standard across terminal
emulators.